Replace the canvas by a simple div
This is much simpler, since we only have to display text,
even with some decorations,and this should handle long
comments much more nicely. Just remains to add one anchor
per comment, and detach them when not wanted any more.

The markers list seems to be the place to store all that.