Generalize the JS code independently of the viewer
No more nasty global variables pervading inside every function.
Well, at least, the viewer is now a parameter of most functions,
and this could also be done with the list of comments. These are
still global in the view, but now the code is generic, and it
shall become possible to experiment with multiple viewers on
the page.