Better rendering of the comments
The goal in this commit is more to leave the possibility to
make the comments directly editable inside their divs. The
overlays are nice, and editable, but I could not make them
work with two different viewers. So I propose to add the
functions to edit and delete the comments directly inside
their column, which would make the code reusable when
bumping them.